Split config into two parts: base and full
The base config will be the bare minimum that tooling projects need in
order to lint themselves. It doesn't have settings for react or jest.
The full config includes settings for react and jest plugins.

This commit is needed so that tooling projects don't need to install the
jest and react plugins.